requirements.txt
#DO NOT CHANGE THIS
pillow==10.3.0
PyAutoGui==0.9.54
pytesseract==0.3.10
pyinstaller==6.8.0